using next is fixed for ThreadNext
